Move component to token::legacy module
